Fix typo in "catchall" constant
The original API has "catchall" spelled as "CATACHALL" in the constants in NB_Single.  This corrects the typo, but also retains the original constant to maintain compatibility as removing a constant would cause an E_ERROR for anyone who hasn't updated.

Feel free to remove the old definition if that isn't a concern.